backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Data & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23

# Database
psycopg2-binary==2.9.9

# Security & Auth
PyJWT==2.8.0
bcrypt==5.0.0
python-dotenv==1.0.0

# Image Processing & ML
pillow==10.1.0
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1

# Optional: TensorFlow (can be removed if not needed)
# tensorflow==2.14.0
# keras==2.14.0

# Logging & Monitoring
python-json-logger==2.0.7

# Database Pool Connection (optional, for better performance)
sqlalchemy-utils==0.41.1
